% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcasting.R
\name{broadcast_black_pixel}
\alias{broadcast_black_pixel}
\title{Broadcast Black Pixel Update}
\usage{
broadcast_black_pixel(pub_socket, position, walker_id)
}
\arguments{
\item{pub_socket}{Publisher socket}

\item{position}{Vector c(x, y) of black pixel position}

\item{walker_id}{ID of walker that created the black pixel}
}
\description{
Serializes and broadcasts a black pixel update to all subscribers
}
