% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker_dynamic.R
\name{choose_next_position}
\alias{choose_next_position}
\title{Choose Next Position}
\usage{
choose_next_position(position, grid, neighborhood)
}
\arguments{
\item{position}{Vector c(x, y) current position}

\item{grid}{Matrix current grid state}

\item{neighborhood}{Character "4-hood" or "8-hood"}
}
\value{
Vector c(x, y) next position
}
\description{
Randomly selects next position from valid neighbors
}
\keyword{internal}
