% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{count_black_pixels}
\alias{count_black_pixels}
\title{Count Black Pixels in Grid}
\usage{
count_black_pixels(grid)
}
\arguments{
\item{grid}{Numeric matrix. The simulation grid.}
}
\value{
Integer. Number of black pixels.
}
\description{
Count Black Pixels in Grid
}
\examples{
grid <- initialize_grid(10)
count_black_pixels(grid)  # 1 (only center)

}
