% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{generate_walker_positions}
\alias{generate_walker_positions}
\title{Generate Random Starting Positions for Walkers}
\usage{
generate_walker_positions(n_walkers, grid, avoid_black = TRUE)
}
\arguments{
\item{n_walkers}{Integer. Number of walkers to create.}

\item{grid}{Numeric matrix. The simulation grid.}

\item{avoid_black}{Logical. If TRUE, avoids placing walkers on black pixels.}
}
\value{
A list of integer vectors, each of length 2 (row, col).
}
\description{
Creates random starting positions for walkers, avoiding the center pixel
and any black pixels.
}
\examples{
grid <- initialize_grid(10)
positions <- generate_walker_positions(5, grid)

}
