% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{get_black_percentage}
\alias{get_black_percentage}
\title{Get Percentage of Black Pixels}
\usage{
get_black_percentage(grid)
}
\arguments{
\item{grid}{Numeric matrix. The simulation grid.}
}
\value{
Numeric. Percentage of black pixels (0-100).
}
\description{
Get Percentage of Black Pixels
}
\examples{
grid <- initialize_grid(10)
get_black_percentage(grid)  # 1\% for 10x10 grid

}
