% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R, R/walker_dynamic.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighboring Positions}
\usage{
get_neighbors(position, grid_size, neighborhood)

get_neighbors(position, grid_size, neighborhood)
}
\arguments{
\item{position}{Vector c(x, y) current position}

\item{grid_size}{Integer grid dimensions}

\item{neighborhood}{Character "4-hood" or "8-hood"}

\item{pos}{Integer vector of length 2 (row, col).}
}
\value{
A list of integer vectors, each of length 2.

List of neighbor positions
}
\description{
Returns all valid neighbor positions for a given position.

Returns list of neighbor positions based on neighborhood type
}
\examples{
get_neighbors(c(5, 5), "4-hood")  # Returns 4 neighbors
get_neighbors(c(5, 5), "8-hood")  # Returns 8 neighbors

}
\keyword{internal}
