% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker_dynamic.R
\name{get_neighbors_bounded}
\alias{get_neighbors_bounded}
\title{Get Neighbor Positions}
\usage{
get_neighbors_bounded(position, grid_size, neighborhood)
}
\arguments{
\item{position}{Vector c(x, y) current position}

\item{grid_size}{Integer grid dimensions}

\item{neighborhood}{Character "4-hood" or "8-hood"}
}
\value{
List of neighbor positions
}
\description{
Returns list of neighbor positions based on neighborhood type
}
\keyword{internal}
