% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker_dynamic.R
\name{handle_boundary}
\alias{handle_boundary}
\title{Handle Boundary Condition}
\usage{
handle_boundary(position, boundary, nrow, ncol)
}
\arguments{
\item{position}{Vector c(x, y) position that hit boundary}

\item{boundary}{Character "wrap" or "terminate"}

\item{nrow}{Integer grid rows}

\item{ncol}{Integer grid columns}
}
\value{
Vector c(x, y) adjusted position
}
\description{
Handle Boundary Condition
}
\keyword{internal}
