% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcasting.R
\name{init_publisher_socket}
\alias{init_publisher_socket}
\title{Initialize Publisher Socket}
\usage{
init_publisher_socket(port = 5555)
}
\arguments{
\item{port}{Port number for socket (default 5555)}
}
\value{
Publisher socket object
}
\description{
Creates a nanonext publisher socket for broadcasting black pixel updates
}
