% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcasting.R
\name{init_subscriber_socket}
\alias{init_subscriber_socket}
\title{Initialize Subscriber Socket}
\usage{
init_subscriber_socket(host = "localhost", port = 5555)
}
\arguments{
\item{host}{Host address (default "localhost")}

\item{port}{Port number (default 5555)}
}
\value{
Subscriber socket object
}
\description{
Creates a nanonext subscriber socket for receiving black pixel updates
}
