% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Initialize a Grid}
\usage{
initialize_grid(n, center_black = TRUE)
}
\arguments{
\item{n}{Integer. The size of the grid (n x n). Must be >= 3.}

\item{center_black}{Logical. If TRUE, initializes the center pixel as black.
Default is TRUE.}
}
\value{
A numeric matrix of size n x n with 0 (white) and 1 (black) values.
}
\description{
Creates an n x n grid for the random walk simulation. By default, the center
pixel is set to black (1), and all other pixels are white (0).
}
\examples{
grid <- initialize_grid(10)
grid[5, 5]  # Center pixel should be 1 (black)

}
