% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker_dynamic.R
\name{is_boundary}
\alias{is_boundary}
\title{Check if Position is at Boundary}
\usage{
is_boundary(position, nrow, ncol)
}
\arguments{
\item{position}{Vector c(x, y)}

\item{nrow}{Integer number of rows}

\item{ncol}{Integer number of columns}
}
\value{
Logical TRUE if at boundary
}
\description{
Check if Position is at Boundary
}
\keyword{internal}
