% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{print_simulation_result}
\alias{print_simulation_result}
\title{Print Simulation Result}
\usage{
print_simulation_result(result)
}
\arguments{
\item{result}{List. Result from run_simulation().}
}
\description{
Print Simulation Result
}
