% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker_dynamic.R
\name{sample_start_position}
\alias{sample_start_position}
\title{Sample Random Starting Position}
\usage{
sample_start_position(nrow, ncol)
}
\arguments{
\item{nrow}{Integer grid rows}

\item{ncol}{Integer grid columns}
}
\value{
Vector c(x, y) random position
}
\description{
Sample Random Starting Position
}
\keyword{internal}
