% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{set_pixel_black}
\alias{set_pixel_black}
\title{Set Pixel Value in Grid}
\usage{
set_pixel_black(grid, pos, boundary = "terminate")
}
\arguments{
\item{grid}{Numeric matrix. The simulation grid.}

\item{pos}{Integer vector of length 2 (row, col).}

\item{boundary}{Character. Either "terminate" or "wrap". Default is "terminate".}
}
\value{
Modified grid matrix.
}
\description{
Sets the value at a given position to black (1).
}
\examples{
grid <- initialize_grid(10)
grid <- set_pixel_black(grid, c(3, 3))
grid[3, 3]  # 1

}
