% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{sim_input_ui}
\alias{sim_input_ui}
\title{Simulation Input Module UI}
\usage{
sim_input_ui(id)
}
\arguments{
\item{id}{Character string. Module namespace ID.}
}
\value{
A shiny tagList containing the input UI elements.
}
\description{
Creates the UI for simulation input parameters.
}
