% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{step_walker}
\alias{step_walker}
\title{Move Walker One Step}
\usage{
step_walker(walker, neighborhood = "4-hood", boundary = "terminate")
}
\arguments{
\item{walker}{List. Walker object.}

\item{neighborhood}{Character. "4-hood" or "8-hood".}

\item{boundary}{Character. "terminate" or "wrap".}
}
\value{
Modified walker object.
}
\description{
Moves the walker one random step in the neighborhood.
}
