% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{wrap_position}
\alias{wrap_position}
\title{Wrap Position Around Grid Boundaries (Torus Topology)}
\usage{
wrap_position(pos, n)
}
\arguments{
\item{pos}{Integer vector of length 2 (row, col).}

\item{n}{Integer. Grid size.}
}
\value{
Integer vector of length 2 with wrapped coordinates.
}
\description{
Wrap Position Around Grid Boundaries (Torus Topology)
}
\examples{
wrap_position(c(0, 5), 10)   # c(10, 5)
wrap_position(c(11, 5), 10)  # c(1, 5)
wrap_position(c(5, 0), 10)   # c(5, 10)

}
