% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighboring Positions}
\usage{
get_neighbors(pos, neighborhood = c("4-hood", "8-hood"))
}
\arguments{
\item{pos}{Integer vector of length 2 (row, col).}

\item{neighborhood}{Character. Either "4-hood" (NSEW) or "8-hood" (includes diagonals).}
}
\value{
A list of integer vectors, each of length 2.
}
\description{
Returns all valid neighbor positions for a given position.
}
\examples{
get_neighbors(c(5, 5), "4-hood")
get_neighbors(c(5, 5), "8-hood")

}
