% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{has_black_neighbor}
\alias{has_black_neighbor}
\title{Check if Walker Has Black Neighbor}
\usage{
has_black_neighbor(
  walker,
  grid,
  neighborhood = "4-hood",
  boundary = "terminate"
)
}
\arguments{
\item{walker}{List. Walker object.}

\item{grid}{Numeric matrix. The simulation grid.}

\item{neighborhood}{Character. "4-hood" or "8-hood".}

\item{boundary}{Character. Boundary condition.}
}
\value{
Logical. TRUE if walker has at least one black neighbor.
}
\description{
Checks if any of the walker's neighbors are black pixels.
}
